requirements.txt
pytest~=7.3.1
requests~=2.28.2
ratelimiter~=1.2.0.post0
SQLAlchemy~=2.0.9
PyMySQL~=1.0.3
discord.py~=2.2.2
tzdata~=2023.3